Byron.Spec.Ledger.Core

Documentation

newtype Hash

An encoded hash of part of the system.

Nothing is used to signal to the elaborators (i.e. the algorithms that translate abstract data into data concrete) that they should produce an invalid concrete hash.

Constructors

Hash

Fields

unHash :: Size

Lens' DSState [(Slot, (VKeyGenesis, VKey))]

type Rep VKeyGenesis

Instance details

Defined in Byron.Spec.Ledger.Core

type Rep VKeyGenesis = D1 ('MetaData "VKeyGenesis" "Byron.Spec.Ledger.Core" "byron-spec-ledger-1.0.1.0-inplace" 'True) (C1 ('MetaCons "VKeyGenesis" 'PrefixI 'True) (S1 ('MetaSel ('Just "unVKeyGenesis") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 VKey)))

mkVKeyGenesis :: Natural -> VKeyGenesis

mkVkGenesisSet

Arguments

:: Word8
Number of genesis keys

-> Set VKeyGenesis

Make a set of genesis keys. The genesis keys are continuously numbered from 0 to the given number of genesis keys minus 1.

data KeyPair

Key Pair.

Constructors

KeyPair

Fields

sKey :: SKey

vKey :: VKey

type Rep KeyPair

Instance details

Defined in Byron.Spec.Ledger.Core

type Rep KeyPair = D1 ('MetaData "KeyPair" "Byron.Spec.Ledger.Core" "byron-spec-ledger-1.0.1.0-inplace" 'False) (C1 ('MetaCons "KeyPair" 'PrefixI 'True) (S1 ('MetaSel ('Just "sKey") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedStrict) (Rec0 SKey) :*: S1 ('MetaSel ('Just "vKey") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedStrict) (Rec0 VKey)))

keyPair :: Owner -> KeyPair

Return a key pair for a given owner.

data Sig a

A digital signature.

Constructors

Sig a Owner

Lens' DSState [(Slot, (VKeyGenesis, VKey))]

type Rep Slot

Instance details

Defined in Byron.Spec.Ledger.Core

type Rep Slot = D1 ('MetaData "Slot" "Byron.Spec.Ledger.Core" "byron-spec-ledger-1.0.1.0-inplace" 'True) (C1 ('MetaCons "Slot" 'PrefixI 'True) (S1 ('MetaSel ('Just "unSlot") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Word64)))

newtype SlotCount

A number of slots.

We use this newtype to distinguish between a cardinal slot and a relative period of slots, and also to distinguish between number of slots and number of blocks.

Constructors

SlotCount

Fields

unSlotCount :: Size

ThunkInfo)
showTypeOf :: Proxy SlotCount -> String

type Rep SlotCount

Instance details

Defined in Byron.Spec.Ledger.Core

type Rep SlotCount = D1 ('MetaData "SlotCount" "Byron.Spec.Ledger.Core" "byron-spec-ledger-1.0.1.0-inplace" 'True) (C1 ('MetaCons "SlotCount" 'PrefixI 'True) (S1 ('MetaSel ('Just "unSlotCount") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Word64)))

addSlot :: Slot -> SlotCount -> Slot

Add a slot count to a slot.

(+.) :: Slot -> SlotCount -> Slot infixl 6

An alias for addSlot

minusSlot :: Slot -> SlotCount -> Slot

Subtract a slot count from a slot.

This is bounded below by 0.

(-.) :: Slot -> SlotCount -> Slot infixl 6

An alias for minusSlot

(*.) :: Word64 -> BlockCount -> SlotCount infixl 7

Multiply the block count by the given constant. This function does not check for overflow.

minusSlotMaybe :: Slot -> SlotCount -> Maybe Slot

Subtract a slot count from a slot.

In case the slot count is greater than the slot's index, it returns Nothing.

newtype BlockCount Size

ThunkInfo)
showTypeOf :: Proxy Lovelace -> String

type Rep Lovelace

Instance details

Defined in Byron.Spec.Ledger.Core

type Rep Lovelace = D1 ('MetaData "Lovelace" "Byron.Spec.Ledger.Core" "byron-spec-ledger-1.0.1.0-inplace" 'True) (C1 ('MetaCons "Lovelace" 'PrefixI 'True) (S1 ('MetaSel ('Just "unLovelace") 'NoSourceUnpackedness 'NoSourceStrictness 'DecidedLazy) (Rec0 Integer)))

lovelaceCap :: Lovelace

Constant amount of Lovelace in the system.

class Relation m where

Minimal complete definition

singleton, dom, range, (◁), (⋪), (▷), (⋫), (∪), (⨃), (<=◁), (▷<=), (▷>=), size

Associated Types

type Domain m :: Type

type Range m :: Type

Methods

singleton :: Domain m -> Range m -> m

dom :: Ord (Domain m) => m -> Set (Domain m)

Domain

range :: Ord (Range m) => m -> Set (Range m)

Range

(◁) :: (Ord (Domain m), Foldable f) => f (Domain m) -> m -> m

Domain restriction

Unicode: 25c1

(<|) :: (Ord (Domain m), Foldable f) => f (Domain m) -> m -> m

Domain restriction

Unicode: 25c1

(⋪) :: (Ord (Domain m), Foldable f) => f (Domain m) -> m -> m

Domain exclusion

Unicode: 22ea

(</|) :: (Ord (Domain m), Foldable f) => f (Domain m) -> m -> m

Domain exclusion

Unicode: 22ea

(▷) :: Ord (Range m) => m -> Set (Range m) -> m

Range restriction

Unicode: 25b7

(|>) :: Ord (Range m) => m -> Set (Range m) -> m

Range restriction

Unicode: 25b7

(⋫) :: Ord (Range m) => m -> Set (Range m) -> m

Range exclusion

Unicode: 22eb

(|/>) :: Ord (Range m) => m -> Set (Range m) -> m

Range exclusion

Unicode: 22eb

(∪) :: (Ord (Domain m), Ord (Range m)) => m -> m -> m

Union

(⨃) :: (Ord (Domain m), Ord (Range m), Foldable f) => m -> f (Domain m, Range m) -> m

Union Override Right

(<=◁) :: Ord (Domain m) => Domain m -> m -> m infixl 5

Restrict domain to values less or equal than the given value.

Unicode: 25c1

(▷<=) :: Ord (Range m) => m -> Range m -> m infixl 5

Restrict range to values less or equal than the given value

Unicode: 25b7

(▷>=) :: Ord (Range m) => m -> Range m -> m infixl 5

Restrict range to values greater or equal than the given value

Unicode: 25b7

size :: Integral n => m -> n

Size of the relation

(∈) :: (Eq a, Foldable f) => a -> f a -> Bool

Alias for elem.

Unicode: 2208

(∉) :: (Eq a, Foldable f) => a -> f a -> Bool infixl 4

Alias for not elem.

Unicode: 2209

(∪+) :: (Ord a, Ord b, Num b) => Map a b -> Map a b -> Map a b

Union override plus is (AB)∪(BA)∪{k|->v1+v2 | k|->v1 : A / k|->v2 : B}

(⊆) :: (Foldable f, Foldable g, Ord a) => f a -> g a -> Bool

Inclusion among foldables.

Unicode: 2286

toSet :: (Foldable f, Ord a) => f a -> Set a

(∩) :: Ord a => Set a -> Set a -> Set a