Byron.Spec.Ledger.Delegation

Delegation scheduling

data SDELEG

Delegation scheduling rules

data SDELEGS

Delegation scheduling sequencing

data DSState

Delegation scheduling state

Constructors

DSState [(Slot, (VKeyGenesis, VKey))] (Set (Epoch, VKeyGenesis))

_dSStateScheduledDelegations :: DSState -> [(Slot, (VKeyGenesis, VKey))]

_dSStateKeyEpochDelegations :: DSState -> Set (Epoch, VKeyGenesis)

data DCert

A delegation certificate.

Constructors

DCert VKeyGenesis VKey Epoch (Sig (VKey, Epoch))

delegator :: DCert -> VKeyGenesis

Key that delegates

delegate :: DCert -> VKey

Key that the delegator is delegating to

depoch :: DCert -> Epoch

Certificate epoch

dwho :: DCert -> (VKeyGenesis, VKey)

Who is delegating to whom.

mkDCert :: VKeyGenesis -> Sig (VKey, Epoch) -> VKey -> Epoch -> DCert

signature :: DCert -> Sig (VKey, Epoch)

Witness for the delegation certificate

Delegation activation

data ADELEG

Delegation rules

data ADELEGS

Delegation rules sequencing

data DSEnv

Delegation scheduling environment

Constructors

DSEnv

Fields

_dSEnvAllowedDelegators :: Set VKeyGenesis

Set of allowed delegators

_dSEnvEpoch :: Epoch

Current epoch

_dSEnvSlot :: Slot

Current slot

_dSEnvK :: BlockCount

Chain stability parameter

allowedDelegators :: HasAllowedDelegators s a => Lens' s a

data DState

Delegation state

Constructors

DState

Fields

_dStateDelegationMap :: Bimap VKeyGenesis VKey

_dStateLastDelegation :: Map VKeyGenesis Slot

When was the last time each genesis key delegated.

Delegation interface

data DELEG

Delegation interface

type DIEnv = DSEnv

Interface environment is the same as scheduling environment.

data DIState

Constructors

DIState (Bimap VKeyGenesis VKey) (Map VKeyGenesis Slot) [(Slot, (VKeyGenesis, VKey))] (Set (Epoch, VKeyGenesis))

_dIStateDelegationMap :: DIState -> Bimap VKeyGenesis VKey

_dIStateLastDelegation :: DIState -> Map VKeyGenesis Slot

_dIStateScheduledDelegations :: DIState -> [(Slot, (VKeyGenesis, VKey))]

_dIStateKeyEpochDelegations :: DIState -> Set (Epoch, VKeyGenesis)

liveAfter :: BlockCount -> SlotCount

Compute after which slot the delegation certificate will become live, using the chain stability parameter.

data EpochDiff

Constructors

EpochDiff

Fields

currentEpoch :: Epoch

certEpoch :: Epoch

State lens fields

slot :: HasSlot s a => Lens' s a

epoch :: HasEpoch s a => Lens' s a

delegationMap :: HasDelegationMap s a => Lens' s a

State lens type classes

class HasScheduledDelegations s a | s -> a

Minimal complete definition

scheduledDelegations

scheduledDelegations :: HasScheduledDelegations s a => Lens' s a

dmsL :: HasDelegationMap a (Bimap VKeyGenesis VKey) => Lens' a (Bimap VKeyGenesis VKey)

Generators

dcertGen :: DSEnv -> Set (Epoch, VKeyGenesis) -> Gen (Maybe DCert)

dcertsGen :: DSEnv -> DIState -> Gen [DCert]

initialEnvFromGenesisKeys

Arguments

:: Word8
Number of genesis keys.

-> Word64
Chain length

-> Gen DSEnv

Generate an initial DELEG environment from the given number of genesis keys.

randomDCertGen :: Environment DELEG -> Gen DCert

Generate a random delegation certificate, which has a high probability of failing since we do not consider the current delegation state. So for instance, we could generate a delegation certificate for a genesis key that already delegated in this epoch.

goblinGensDELEG :: [SignalGenerator DELEG]

Functions on delegation state

delegatorOf :: Bimap VKeyGenesis VKey -> VKey -> Maybe VKeyGenesis

Support Functions for delegation properties

delegatorDelegate :: DCert -> (VKeyGenesis, VKey)

emptyDelegationPayloadRatio :: [[DCert]] -> Double

Ratio of certificate groups that are empty

thisEpochDelegationsRatio :: [(Epoch, Epoch)] -> Double

Ratio of certificates that delegate to _this_ epoch, where each certificate is represented by (current epoch,cert epoch)

nextEpochDelegationsRatio :: [(Epoch, Epoch)] -> Double

Ratio of certificates that delegate to the _next_ epoch, where each certificate is represented by (current epoch,cert epoch)

selfDelegationsRatio :: [DCert] -> Double

Ratio of certificates that "delegate to self", that is, where the delegator and delegate are the same

multipleDelegationsRatio :: [DCert] -> Double

Ratio of delegates that have multiple delegators that are delegating to them

maxDelegationsTo :: [DCert] -> Int

The maximum number of delegators to any particular delegate

changedDelegationsRatio :: [DCert] -> Double

Ratio of delegators that have changed their delegations

maxChangedDelegations :: [DCert] -> Int

The maximum number of change-of-delegate for any particular delegator

repeatedDelegationsRatio :: [DCert] -> Double

Ratio of repeated delegations to all delegations

maxRepeatedDelegations :: [DCert] -> Int

The maximum number of repeated delegations in the given certificates

maxCertsPerBlock :: [[DCert]] -> Int

Predicate failures

data AdelegPredicateFailure

None of these PredicateFailures are actually "throwable". The disjuction of the rules' preconditions is True, which means that one of them will pass. The PredicateFailure just act as switches to direct control flow to the successful one.

Constructors

S_BeforeExistingDelegation

S_NoLastDelegation

S_AfterExistingDelegation

S_AlreadyADelegateOf VKey VKeyGenesis

data AdelegsPredicateFailure

Constructors

ADelegFailure (PredicateFailure ADELEG)

data SdelegPredicateFailure

These PredicateFailures are all "throwable". The disjunction of the rules' preconditions is not True - the PredicateFailures represent False cases.

Constructors

IsNotGenesisKey

EpochInThePast EpochDiff

EpochPastNextEpoch EpochDiff

HasAlreadyDelegated

IsAlreadyScheduled

DoesNotVerify

data SdelegsPredicateFailure

Constructors

SDelegFailure (PredicateFailure SDELEG)

data MsdelegPredicateFailure

Constructors

SDELEGFailure (PredicateFailure SDELEG)

data DelegPredicateFailure

Constructors

SDelegSFailure (PredicateFailure SDELEGS)

ADelegSFailure (PredicateFailure ADELEGS)

tamperedDcerts :: DIEnv -> DIState -> Gen [DCert]