Byron.Spec.Ledger.GlobalParams

Description

Ledger global parameters.

Documentation

epochFirstSlot ∷ BlockCount → Epoch → Slot

Given the chain stability parameter, calculate the first slot in a given epoch.

lovelaceCap ∷ Lovelace

Constant amount of Lovelace in the system.

slotsPerEpoch ∷ Integral n ⇒ BlockCount → n

Given the chain stability parameter, often referred to as k, which is expressed in an amount of blocks, return the number of slots contained in an epoch.

slotsPerEpochToK ∷ Integral n ⇒ n → BlockCount

The inverse of slotsPerEpoch: given a number of slots per-epoch, return the chain stability parameter k.

c ∷ Word64

Factor used to bound the concrete size by the abstract size.

This constant should satisfy that given an elaboration function elaborate which elaborates abstract values intro concrete ones, for each abstract data value a we have:

size (elaborate a) <= c * abstractSize a

TODO: we need to investigate what this factor is, and probably use different factors for different data types (update, UTxO transactions, etc).