Byron.Spec.Ledger.UTxO.Generators

genList :: NodeT m [a])

Temporarily defined here until hedgehog exposes this function

genTxFromUTxO

Arguments

:: [Addr]
List of addresses to choose from as recipients of the transaction outputs.

-> (Tx -> Lovelace)
Fee calculation

-> UTxO
UTxO used to determine which unspent outputs can be used in the transaction.

-> Property

A property to test that the entire shrink tree generated by genInputOutput maintains the invariant that the inputs and outputs have equal sums

NB: This uses small values for the list and values, because we force the entire shrink tree, which can grow very large

genInputOutput

Arguments

:: [input]
List of input to take a subsequence of

-> (input -> Gen ([input], [output])

Generate a subsequence of the inputs and split the value into a number of outputs

The shrink tree maintains the invariant that the sum of the inputs is equal to the sum of the outputs. This is generalised to any type of input where we can view the contained value, and type of output where we can generate the type from a value, and view and modify the contained value

interleaveInputOutputTreeT :: TreeT m ([input], [output])]

Shrink the left value of a tuple, preserving the total value stored in the right value

Used as part of genInputOutput, so see there for details of the arguments

subFromList

Arguments

:: (Gen [a]

Given an input value and functions to generate, view, and update some type a based on that value, split the input into a number of as, preserving the value throughout the shrink tree

interleaveTreeTPreserving :: TreeT m [a]]

Drop one of the outputs, preserving the invariant by moving its value to the left

Used as part of genSplitValue, so see there for details of the arguments

shrinkOnePreserving :: TreeT m [a]]

Shrink a value in a list, preserving the total value by moving the lost value to the left

Used as part of genSplitValue, so see there for details of the arguments

viewTwo :: [a] -> [([a], a, a, [a])]

All the ways of choosing two consecutive values from a list