Byron.Spec.Ledger.Core

Documentation

newtype Hash

An encoded hash of part of the system.

Nothing is used to signal to the elaborators (i.e. the algorithms that translate abstract data into data concrete) that they should produce an invalid concrete hash.

Constructors

Hash

Fields

unHash ∷ Maybe Int

isValid ∷ Hash → Bool

class HasHash a where

Hash part of the ledger payload

Methods

hash ∷ a → Hash

newtype Owner

Representation of the owner of key pair.

Constructors

Owner

Fields

unOwner ∷ Natural

class HasOwner a where

Methods

owner ∷ a → Owner

newtype SKey

Signing Key.

Constructors

SKey Owner

newtype VKey

Verification Key.

Constructors

VKey Owner

newtype VKeyGenesis

A genesis key is a specialisation of a generic VKey.

Constructors

VKeyGenesis

Fields

unVKeyGenesis ∷ VKey

mkVKeyGenesis ∷ Natural → VKeyGenesis

mkVkGenesisSet

Arguments

∷ Word8
Number of genesis keys

→ Set VKeyGenesis

Make a set of genesis keys. The genesis keys are continuously numbered from 0 to the given number of genesis keys minus 1.

data KeyPair

Key Pair.

Constructors

KeyPair

Fields

sKey ∷ SKey

vKey ∷ VKey

keyPair ∷ Owner → KeyPair

Return a key pair for a given owner.

data Sig a

A digital signature.

Constructors

Sig a Owner

sign ∷ SKey → a → Sig a

Produce a digital signature

verify ∷ Eq a ⇒ VKey → a → Sig a → Bool

Verify a digital signature

newtype Epoch

Constructors

Epoch

Fields

unEpoch ∷ Word64

newtype Slot

Constructors

Slot

Fields

unSlot ∷ Word64

newtype SlotCount

A number of slots.

We use this newtype to distinguish between a cardinal slot and a relative period of slots, and also to distinguish between number of slots and number of blocks.

Constructors

SlotCount

Fields

unSlotCount ∷ Word64

addSlot ∷ Slot → SlotCount → Slot

Add a slot count to a slot.

(+.) ∷ Slot → SlotCount → Slot infixl 6

An alias for addSlot

minusSlot ∷ Slot → SlotCount → Slot

Subtract a slot count from a slot.

This is bounded below by 0.

(-.) ∷ Slot → SlotCount → Slot infixl 6

An alias for minusSlot

(*.) ∷ Word64 → BlockCount → SlotCount infixl 7

Multiply the block count by the given constant. This function does not check for overflow.

minusSlotMaybe ∷ Slot → SlotCount → Maybe Slot

Subtract a slot count from a slot.

In case the slot count is greater than the slot's index, it returns Nothing.

newtype BlockCount

Constructors

BlockCount

Fields

unBlockCount ∷ Word64

newtype Addr

The address of a transaction output, used to identify the owner.

Constructors

Addr VKey

mkAddr ∷ Natural → Addr

Create an address from a number.

newtype Lovelace

A unit of value held by a UTxO.

Constructors

Lovelace

Fields

unLovelace ∷ Integer

lovelaceCap ∷ Lovelace

Constant amount of Lovelace in the system.

class Relation m where

Minimal complete definition

singleton, dom, range, (◁), (⋪), (▷), (⋫), (∪), (⨃), (<=◁), (▷<=), (▷>=), size

Associated Types

type Domain m ∷ Type

type Range m ∷ Type

Methods

singleton ∷ Domain m → Range m → m

dom ∷ Ord (Domain m) ⇒ m → Set (Domain m)

Domain

range ∷ Ord (Range m) ⇒ m → Set (Range m)

Range

(◁) ∷ (Ord (Domain m), Foldable f) ⇒ f (Domain m) → m → m

Domain restriction

Unicode: 25c1

(<|) ∷ (Ord (Domain m), Foldable f) ⇒ f (Domain m) → m → m

Domain restriction

Unicode: 25c1

(⋪) ∷ (Ord (Domain m), Foldable f) ⇒ f (Domain m) → m → m

Domain exclusion

Unicode: 22ea

(</|) ∷ (Ord (Domain m), Foldable f) ⇒ f (Domain m) → m → m

Domain exclusion

Unicode: 22ea

(▷) ∷ Ord (Range m) ⇒ m → Set (Range m) → m

Range restriction

Unicode: 25b7

(|>) ∷ Ord (Range m) ⇒ m → Set (Range m) → m

Range restriction

Unicode: 25b7

(⋫) ∷ Ord (Range m) ⇒ m → Set (Range m) → m

Range exclusion

Unicode: 22eb

(|/>) ∷ Ord (Range m) ⇒ m → Set (Range m) → m

Range exclusion

Unicode: 22eb

(∪) ∷ (Ord (Domain m), Ord (Range m)) ⇒ m → m → m

Union

(⨃) ∷ (Ord (Domain m), Ord (Range m), Foldable f) ⇒ m → f (Domain m, Range m) → m

Union Override Right

(<=◁) ∷ Ord (Domain m) ⇒ Domain m → m → m infixl 5

Restrict domain to values less or equal than the given value.

Unicode: 25c1

(▷<=) ∷ Ord (Range m) ⇒ m → Range m → m infixl 5

Restrict range to values less or equal than the given value

Unicode: 25b7

(▷>=) ∷ Ord (Range m) ⇒ m → Range m → m infixl 5

Restrict range to values greater or equal than the given value

Unicode: 25b7

size ∷ Integral n ⇒ m → n

Size of the relation

(∈) ∷ (Eq a, Foldable f) ⇒ a → f a → Bool

Alias for elem.

Unicode: 2208

(∉) ∷ (Eq a, Foldable f) ⇒ a → f a → Bool infixl 4

Alias for not elem.

Unicode: 2209

(∪+) ∷ (Ord a, Ord b, Num b) ⇒ Map a b → Map a b → Map a b

Union override plus is (AB)∪(BA)∪{k|->v1+v2 | k|->v1 : A / k|->v2 : B}

(⊆) ∷ (Foldable f, Foldable g, Ord a) ⇒ f a → g a → Bool

Inclusion among foldables.

Unicode: 2286

toSet ∷ (Foldable f, Ord a) ⇒ f a → Set a

(∩) ∷ Ord a ⇒ Set a → Set a → Set a