Byron.Spec.Ledger.STS.UTXOW

Description

UTXO transition system with witnessing

Documentation

data UTXOW

data UtxowPredicateFailure

These PredicateFailures are all throwable.

Constructors

UtxoFailure (PredicateFailure UTXO)

InsufficientWitnesses

authTxin ∷ VKey → TxIn → UTxO → Bool

Determine if a UTxO input is authorized by a given key.

witnessed ∷ Tx → UTxO → Bool

Given a ledger state, determine if the UTxO witnesses in a given transaction are sufficient. TODO - should we only check for one witness for each unique input address?

traceAddrs ∷ [Addr]

Constant list of addresses intended to be used in the generators.

coverUtxoFailure

Arguments

∷ ∀ m a. (MonadTest m, HasCallStack, Data a)

⇒ CoverPercentage
Minimum percentage that each failure must occur.

→ a
Structure containing the failures

→ m ()

Check that all the relevant predicate failures are covered.