Byron.Spec.Ledger.STS.UTXOW

Description

UTXO transition system with witnessing

Documentation

data UTXOW

authTxin ∷ VKey → TxIn → UTxO → Bool

Determine if a UTxO input is authorized by a given key.

witnessed ∷ Tx → UTxO → Bool

Given a ledger state, determine if the UTxO witnesses in a given transaction are sufficient. TODO - should we only check for one witness for each unique input address?

traceAddrs ∷ [Addr]

Constant list of addresses intended to be used in the generators.

goblinGensUTXOW ∷ [SignalGenerator UTXOW]

goblinGen_UTXOW_UtxoFailure_NonPositiveOutputs ∷ SignalGenerator UTXOW

goblinData_UTXOW_UtxoFailure_NonPositiveOutputs ∷ GoblinData Bool

goblinGen_UTXOW_UtxoFailure_InputsNotInUTxO ∷ SignalGenerator UTXOW

goblinData_UTXOW_UtxoFailure_InputsNotInUTxO ∷ GoblinData Bool

goblinGen_UTXOW_UtxoFailure_FeeTooLow ∷ SignalGenerator UTXOW

goblinData_UTXOW_UtxoFailure_FeeTooLow ∷ GoblinData Bool

goblinGen_UTXOW_UtxoFailure_EmptyTxOutputs ∷ SignalGenerator UTXOW

goblinData_UTXOW_UtxoFailure_EmptyTxOutputs ∷ GoblinData Bool

goblinGen_UTXOW_UtxoFailure_EmptyTxInputs ∷ SignalGenerator UTXOW

goblinData_UTXOW_UtxoFailure_EmptyTxInputs ∷ GoblinData Bool

goblinGen_UTXOW_InsufficientWitnesses ∷ SignalGenerator UTXOW

goblinData_UTXOW_InsufficientWitnesses ∷ GoblinData Bool

tamperedTxList ∷ UTxOEnv → UTxOState → Gen [Tx]

coverUtxoFailure

Arguments

∷ (MonadTest m, HasCallStack, Data a)

⇒ CoverPercentage
Minimum percentage that each failure must occur.

→ a
Structure containing the failures

→ m ()

Check that all the relevant predicate failures are covered.