Byron.Spec.Ledger.UTxO

Documentation

newtype TxId

A unique ID of a transaction, which is computable from the transaction.

Constructors

TxId

Fields

getTxId ∷ Hash

data TxIn

The input of a UTxO.

TODO - is it okay to use list indices instead of implementing the Ix Type?

Constructors

TxIn TxId Natural

data TxOut

The output of a UTxO.

Constructors

TxOut

Fields

addr ∷ Addr

value ∷ Lovelace

newtype UTxO

The unspent transaction outputs.

Constructors

UTxO

Fields

unUTxO ∷ Map TxIn TxOut

mapUTxOValues ∷ (Lovelace → Lovelace) → UTxO → UTxO

Apply function uniformly across all outputs

addValue ∷ TxOut → Lovelace → TxOut

fromTxOuts ∷ [TxOut] → UTxO

Construct a UTxO from initial TxOuts

data TxBody

A raw transaction

Constructors

TxBody

Fields

inputs ∷ [TxIn]

outputs ∷ [TxOut]

txid ∷ TxBody → TxId

txValue ∷ TxBody → Lovelace

Total value of a transaction.

txins ∷ TxBody → [TxIn]

Compute the UTxO inputs of a transaction.

txouts ∷ TxBody → UTxO

Compute the UTxO outputs of a transaction.

balance ∷ UTxO → Lovelace

Determine the total balance contained in the UTxO.

pcMinFee ∷ PParams → Tx → Lovelace

txsize ∷ Tx → Int

data Wit

Proof/Witness that a transaction is authorized by the given key holder.

Constructors

Wit VKey (Sig TxBody)

data Tx

A fully formed transaction.

Constructors

Tx

Fields

body ∷ TxBody

witnesses ∷ [Wit]

makeWitness ∷ KeyPair → TxBody → Wit

Create a witness for transaction

makeTxWits ∷ UTxO → TxBody → Tx