Byron.Spec.Ledger.UTxO.Generators

Documentation

genInitialTxOuts ∷ [Addr] → Gen [TxOut]

Generate a set of initial TxOuts from a set of Addrs

genLovelace ∷ Gen Lovelace

genTraverseSubsequence ∷ (a → Gen b) → [a] → Gen [b]

Generate a subsequence of a list of values and traverse the subsequence with a generator producer

genList ∷ Range Int → Gen a → Gen [a]

Generate a list using genTraverseSubsequence

interleaveTreeT ∷ Monad m ⇒ [TreeT m a] → m (NodeT m [a])

Temporarily defined here until hedgehog exposes this function

genTxFromUTxO

Arguments

∷ [Addr]
List of addresses to choose from as recipients of the transaction outputs.

→ (Tx → Lovelace)
Fee calculation

→ UTxO
UTxO used to determine which unspent outputs can be used in the transaction.

→ Gen Tx

Generate a valid transaction from a given UTxO

propGenInputOutput ∷ Property

A property to test that the entire shrink tree generated by genInputOutput maintains the invariant that the inputs and outputs have equal sums

NB: This uses small values for the list and values, because we force the entire shrink tree, which can grow very large

genInputOutput

Arguments

∷ [input]
List of input to take a subsequence of

→ (input → Integer)
A view of the value in the input used to preserve the invariant

→ Gen (Integer → output)
A applicative generator for an output given a value

→ (output → Integer)
A view of the value in the output used to preserve the invariant

→ ((Integer → Integer) → output → output)
An update function for the output value used to preserve the invariant

→ Gen ([input], [output])

Generate a subsequence of the inputs and split the value into a number of outputs

The shrink tree maintains the invariant that the sum of the inputs is equal to the sum of the outputs. This is generalised to any type of input where we can view the contained value, and type of output where we can generate the type from a value, and view and modify the contained value

interleaveInputOutputTreeT ∷ Monad m ⇒ (input → Integer) → (output → Integer) → ((Integer → Integer) → output → output) → (TreeT m [input], TreeT m [output]) → m (NodeT m ([input], [output]))

Used as part of genInputOutput, so see there for details of the arguments

interleaveInputOutput ∷ Monad m ⇒ (input → Integer) → (output → Integer) → ((Integer → Integer) → output → output) → (NodeT m [input], NodeT m [output]) → NodeT m ([input], [output])

Used as part of genInputOutput, so see there for details of the arguments

shrinkRight ∷ Monad m ⇒ (input → Integer) → (output → Integer) → ((Integer → Integer) → output → output) → (NodeT m [input], NodeT m [output]) → [TreeT m ([input], [output])]

Used as part of genInputOutput, so see there for details of the arguments

shrinkLeftPreserving ∷ Monad m ⇒ (input → Integer) → (output → Integer) → ((Integer → Integer) → output → output) → (NodeT m [input], NodeT m [output]) → [TreeT m ([input], [output])]

Shrink the left value of a tuple, preserving the total value stored in the right value

Used as part of genInputOutput, so see there for details of the arguments

subFromList

Arguments

∷ (Num n, Ord n)

⇒ n
The total value to remove from the list

→ (a → n)
A view into the value contained in type a

→ ((n → n) → a → a)
A modifier for the value contained in type a

→ [a]
The list of as to remove value from

→ [a]

Remove total value from a list, removing from the front

propGenSplitValue ∷ Property

A property to check that genSplitValue does indeed preserve the input

genSplitValue

Arguments

∷ Integer
Total value to divide into outputs

→ Gen (Integer → a)
Applicative generator for an output given a value

→ (a → Integer)
A view of the value in the output used to preserve the invariant

→ ((Integer → Integer) → a → a)
A modifier for the value in the output used to preserve the invariant

→ Gen [a]

Given an input value and functions to generate, view, and update some type a based on that value, split the input into a number of as, preserving the value throughout the shrink tree

interleaveTreeTPreserving ∷ Monad m ⇒ (a → Integer) → ((Integer → Integer) → a → a) → [TreeT m a] → m (NodeT m [a])

Used as part of genSplitValue, so see there for details of the arguments

interleavePreserving ∷ Monad m ⇒ (a → Integer) → ((Integer → Integer) → a → a) → [NodeT m a] → NodeT m [a]

Used as part of genSplitValue, so see there for details of the arguments

dropOnePreserving ∷ Monad m ⇒ (a → Integer) → ((Integer → Integer) → a → a) → [NodeT m a] → [TreeT m [a]]

Drop one of the outputs, preserving the invariant by moving its value to the left

Used as part of genSplitValue, so see there for details of the arguments

shrinkOnePreserving ∷ Monad m ⇒ (a → Integer) → ((Integer → Integer) → a → a) → [NodeT m a] → [TreeT m [a]]

Shrink a value in a list, preserving the total value by moving the lost value to the left

Used as part of genSplitValue, so see there for details of the arguments

viewTwo ∷ [a] → [([a], a, a, [a])]

All the ways of choosing two consecutive values from a list