Byron.Spec.Ledger.Update

Documentation

data PParams

Protocol parameters.

Constructors

PParams

Fields

_maxBkSz ∷ !Natural

Maximum (abstract) block size in words

_maxHdrSz ∷ !Natural

Maximum (abstract) block header size in words

_maxTxSz ∷ !Natural

Maximum (abstract) transaction size in words

_maxPropSz ∷ !Natural

Maximum (abstract) update proposal size in words

_bkSgnCntT ∷ !BkSgnCntT

Fraction [0, 1] of the blocks that can be signed by any given key in a window of lenght _bkSgnCntW. This value will be typically between 1/5 and 1/4

_bkSlotsPerEpoch ∷ !SlotCount

Number of slots in an epoch. TODO: this should be removed since the number of slots per epoch should remain constant.

_upTtl ∷ !SlotCount

Update proposal TTL in slots

_scriptVersion ∷ !Natural

Script version

_upAdptThd ∷ !UpAdptThd

Update adoption threshold: a proportion of block issuers that have to endorse a given version to become candidate for adoption

_factorA ∷ !FactorA

Minimum fees per transaction

_factorB ∷ !FactorB

Additional fees per transaction size

newtype BkSgnCntT

Constructors

BkSgnCntT Double

newtype UpAdptThd

Constructors

UpAdptThd Double

newtype FactorB

Constructors

FactorB Int

newtype FactorA

Constructors

FactorA Int

data ProtVer

Protocol version

Constructors

ProtVer

Fields

_pvMaj ∷ Natural

_pvMin ∷ Natural

_pvAlt ∷ Natural

newtype UpId

Constructors

UpId Int

bkSgnCntT ∷ Lens' PParams BkSgnCntT

bkSlotsPerEpoch ∷ Lens' PParams SlotCount

factorA ∷ Lens' PParams FactorA

factorB ∷ Lens' PParams FactorB

maxBkSz ∷ Lens' PParams Natural

maxHdrSz ∷ Lens' PParams Natural

maxPropSz ∷ Lens' PParams Natural

maxTxSz ∷ Lens' PParams Natural

scriptVersion ∷ Lens' PParams Natural

upAdptThd ∷ Lens' PParams UpAdptThd

upTtl ∷ Lens' PParams SlotCount

data SwVer

Constructors

SwVer

Fields

_svName ∷ ApName

_svVer ∷ ApVer

newtype ApVer

Application version

Constructors

ApVer Natural

newtype ApName

Constructors

ApName String

pvAlt ∷ Lens' ProtVer Natural

pvMaj ∷ Lens' ProtVer Natural

pvMin ∷ Lens' ProtVer Natural

data UProp

Update proposal

Constructors

UProp

Fields

_upId ∷ UpId

_upIssuer ∷ VKey

_upParams ∷ PParams

_upPV ∷ ProtVer

_upSwVer ∷ SwVer

_upSig ∷ Sig UpSD

_upSTags ∷ Set STag

System tags involved in the update proposal.

_upMdt ∷ Metadata

Metadata required for performing software updates.

data Metadata

For now we do not have any requirements on metadata.

Constructors

Metadata

type STag = String

System tag, this represents a target operating system for the update (e.g. linux, win64, or mac32).

type UpSD = (ProtVer, PParams, SwVer, Set STag, Metadata)

Part of the update proposal which must be signed

svName ∷ Lens' SwVer ApName

svVer ∷ Lens' SwVer ApVer

data Vote

Constructors

Vote

Fields

_vCaster ∷ VKey

_vPropId ∷ UpId

_vSig ∷ Sig UpId

data UpregPredicateFailure

These PredicateFailures are all throwable.

Constructors

UPVFailure (PredicateFailure UPV)

NotGenesisDelegate

DoesNotVerify

data UPREG

data UpvPredicateFailure

These PredicateFailures are all throwable.

Constructors

UPPVVFailure (PredicateFailure UPPVV)

UPSVVFailure (PredicateFailure UPSVV)

AVChangedInPVUpdate ApName ApVer (Maybe (ApVer, Slot, Metadata))

ParamsChangedInSVUpdate

PVChangedInSVUpdate

data UPV

Update proposal validity

data UppvvPredicateFailure

These PredicateFailures are all "throwable". The disjunction of the rules' preconditions is not True - the PredicateFailures represent False cases.

Constructors

CannotFollowPv

CannotUpdatePv [UpdateConstraintViolation]

AlreadyProposedPv

data UPPVV

data UpsvvPredicateFailure

These PredicateFailures are all "throwable". The disjunction of the rules' preconditions is not True - the PredicateFailures represent False cases.

Constructors

AlreadyProposedSv

CannotFollowSv

InvalidApplicationName

InvalidSystemTags

data UPSVV

Update Proposal Software Version Validation

data UpdateConstraintViolation

Violations on the constraints of the allowed values for new protocol parameters.

Constructors

BlockSizeTooLarge Natural (Threshold Natural)

TransactionSizeTooLarge Natural (Threshold Natural)

ScriptVersionTooLarge Natural (Threshold Natural)

ScriptVersionTooSmall Natural (Threshold Natural)

upId ∷ Lens' UProp UpId

upIssuer ∷ Lens' UProp VKey

upMdt ∷ Lens' UProp Metadata

upPV ∷ Lens' UProp ProtVer

upParams ∷ Lens' UProp PParams

upSTags ∷ Lens' UProp (Set STag)

upSig ∷ Lens' UProp (Sig UpSD)

upSwVer ∷ Lens' UProp SwVer

upSigData ∷ Lens' UProp UpSD

getUpSigData ∷ UProp → UpSD

mkUProp ∷ UpId → VKey → ProtVer → PParams → SwVer → Set STag → Metadata → UProp

inMap ∷ (Ord key, Eq v) ⇒ key → v → Map key v → Bool

Test if a pair is present in a map.

invertMap ∷ (Ord k, Ord v) ⇒ Map k v → Map v (Set k)

Invert a map

Examples:

>>> import qualified Data.Map.Strict as Map
>>> import Byron.Spec.Ledger.Update (invertMap)
>>> invertMap (Map.fromList [('a', 1 :: Int), ('b', 2), ('c', 3), ('d', 1)])
fromList [(1,fromList "ad"),(2,fromList "b"),(3,fromList "c")]

invertBijection ∷ Ord v ⇒ Map k v → Map v k

Invert a map which we assert to be a bijection. If this map is not a bijection, the behaviour is not guaranteed.

Examples:

>>> import qualified Data.Map.Strict as Map
>>> invertBijection (Map.fromList [('a', 1 :: Int), ('b', 2), ('c', 3)])
fromList [(1,'a'),(2,'b'),(3,'c')]

(==>) ∷ Bool → Bool → Bool infix 1

pvCanFollow

Arguments

∷ ProtVer
Next protocol version

→ ProtVer
Previous protocol version

→ Bool

Check whether a protocol version can follow the current protocol version.

checkUpdateConstraints ∷ PParams → UProp → [UpdateConstraintViolation]

Check whether an update proposal marks a valid update

(<=?) ∷ Ord a ⇒ a → a → Maybe (a, Threshold a) infix 4

orError ∷ Maybe (a, b) → (a → b → e) → Maybe e

canUpdate ∷ PParams → UProp → Rule UPPVV ctx ()

svCanFollow ∷ Map ApName (ApVer, Slot, Metadata) → (ApName, ApVer) → Bool

data UpiecPredicateFailure

Constructors

PVBUMPFailure (PredicateFailure PVBUMP)

data UPIEC

data PvbumpPredicateFailure

Constructors

NoPVBUMPFailure

data PVBUMP

data UpiendPredicateFailure

Constructors

UPENDFailure (PredicateFailure UPEND)

data UPIEND

data UpivotesPredicateFailure

Constructors

ApplyVotesFailure (PredicateFailure APPLYVOTES)

data UPIVOTES

data ApplyVotesPredicateFailure

Constructors

UpivoteFailure (PredicateFailure UPIVOTE)

data APPLYVOTES

data UpivotePredicateFailure

Constructors

UPVOTEFailure (PredicateFailure UPVOTE)

data UPIVOTE

data UpiregPredicateFailure

Constructors

UPREGFailure (PredicateFailure UPREG)

data UPIREG

type UPIState = ((ProtVer, PParams), [(Slot, (ProtVer, PParams))], Map ApName (ApVer, Slot, Metadata), Map UpId (ProtVer, PParams), Map UpId (ApName, ApVer, Metadata), Map UpId Slot, Set (UpId, VKeyGenesis), Set (ProtVer, VKeyGenesis), Map UpId Slot)

The update interface state is shared amongst various rules, so we define it as an alias here.

type UPIEnv = (Slot, Bimap VKeyGenesis VKey, BlockCount, Word8)

The update interface environment is shared amongst various rules, so we define it as an alias here.

data UpendPredicateFailure

S_TryNextRule is a structural PredicateFailure, used to fail from one transition rule to the other. The other PredicateFailures are all throwable.

Constructors

ProtVerUnknown ProtVer

S_TryNextRule

CanAdopt ProtVer

CannotAdopt ProtVer

NotADelegate VKey

UnconfirmedProposal UpId

data UPEND

data FadsPredicateFailure

data FADS

data UpvotePredicateFailure

The 3 non-embedded PredicateFailures here are all structural. The disjuntion of the preconditions is True - one rule either fires or the other does.

Constructors

ADDVOTEFailure (PredicateFailure ADDVOTE)

S_HigherThanThdAndNotAlreadyConfirmed

S_CfmThdNotReached

S_AlreadyConfirmed

data UPVOTE

data AddvotePredicateFailure

These PredicateFailures are all throwable.

Constructors

AVSigDoesNotVerify

NoUpdateProposal UpId

VoteByNonGenesisDelegate VKey

RepeatVoteByGenesisDelegate VKey

data ADDVOTE

vCaster ∷ Lens' Vote VKey

vPropId ∷ Lens' Vote UpId

vSig ∷ Lens' Vote (Sig UpId)

mkVote ∷ VKey → UpId → Vote

findKey ∷ (v → Bool) → Map k v → Maybe (k, v)

Find the key that corresponds to the value satisfying the given predicate. In case zero or more than one key is found this function returns Nothing.

delegationMap ∷ UPIEnv → Bimap VKeyGenesis VKey

fstUPIState ∷ UPIState → (ProtVer, PParams)

sndUPIState ∷ UPIState → [(Slot, (ProtVer, PParams))]

trdUPIState ∷ UPIState → Map ApName (ApVer, Slot, Metadata)

emptyUPIState ∷ UPIState

initialPParams ∷ PParams

protocolVersion ∷ UPIState → ProtVer

protocolParameters ∷ UPIState → PParams

applicationVersions ∷ UPIState → Map ApName (ApVer, Slot, Metadata)

confirmedProposals ∷ UPIState → Map UpId Slot

futureAdoptions ∷ UPIState → [(Slot, (ProtVer, PParams))]

endorsements ∷ UPIState → Set (ProtVer, VKeyGenesis)

registeredProtocolUpdateProposals ∷ UPIState → Map UpId (ProtVer, PParams)

upiEnvGen ∷ Gen UPIEnv

dmapGen ∷ Word8 → Gen (Bimap VKeyGenesis VKey)

ppsUpdateFrom ∷ PParams → Gen PParams

Generate a protocol parameter update from a given set of current protocol-parameters, ensuring the consistency of the new protocol parameters w.r.t. the current ones, according to the canUpdate predicate in the formal specification.

TODO: we can move this into a Generator's module, but first we need to disentangle the dependencies. Moving this to Byron.Spec.Ledger.Update.Generators will cause a circular dependency. I think the rules need to be moved into their own modules.

increasingProbabilityAt ∷ Gen a → (a, a) → Gen a

Generate values the given distribution in 90% of the cases, and values at the bounds of the range in 10% of the cases.

This can be used to generate enough extreme values. The exponential and linear distributions provided by hedgehog will generate a small percentage of these (0-1%).

randomUpId ∷ Gen UpId

Generate a random update proposal id, by picking a large number so that the probability of having an update proposal with that id is nearly zero.

reSign ∷ UProp → UProp

Update the signature of the update proposal.

pickHighlyEndorsedProtocolVersion

Arguments

∷ [(ProtVer, Set VKeyGenesis)]
Current set of endorsements

→ Gen (Maybe ProtVer)

Given a list of protocol versions and keys endorsing those versions, generate a protocol-version endorsement, or Nothing if the list of endorsements is empty. The version to be endorsed will be selected from those versions that have the most endorsements.

updateProposalAndVotesGen ∷ UPIEnv → UPIState → Gen (Maybe UProp, [Vote])

Generate an optional update-proposal and a list of votes, given an update environment and state.

The update proposal and votes need to be generated at the same time, since this allow us to generate update votes for update proposals issued in the same block as the votes.

protocolVersionEndorsementGen ∷ UPIEnv → UPIState → Gen ProtVer

Generate an endorsement given an update environment and state.

type family PredicateFailure a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

Orphan instances

Field1 (a, b, c, d, e, f, g, h, i) (a', b, c, d, e, f, g, h, i) a a'

Instance details

Methods

_1 ∷ Lens (a, b, c, d, e, f, g, h, i) (a', b, c, d, e, f, g, h, i) a a'

Field2 (a, b, c, d, e, f, g, h, i) (a, b', c, d, e, f, g, h, i) b b'

Instance details

Methods

_2 ∷ Lens (a, b, c, d, e, f, g, h, i) (a, b', c, d, e, f, g, h, i) b b'

Field3 (a, b, c, d, e, f, g, h, i) (a, b, c', d, e, f, g, h, i) c c'

Instance details

Methods

_3 ∷ Lens (a, b, c, d, e, f, g, h, i) (a, b, c', d, e, f, g, h, i) c c'