Byron.Spec.Ledger.Util

Documentation

mkGoblinGens ∷ String → [String] → Q [Dec]

Take a name (e.g. DELEG) and a list of PredicateFailures in renderPF form (see repo goblins-sts-breeder; STSExtra typeclass) (e.g. [UTXOW_InsufficientWitnesses]) and returns `n+1` declarations, where n is the length of the pfNames list. There will be 1 decl per element of pfNames, which wraps a goblin mutation around the valid SignalGenerator. The final decl is a toplevel list of the previously defined SignalGenerators.